// client/src/utils/styleUtils.tsx
import { css } from 'styled-components';

// Colors
export const light = '#e7f1ff';
export const medium = '#4c8fed';
export const dark = '#144ea2';

export const screenSize = '576px';

export const defaultButtonStyles = css`
  background-color: ${dark};
  color: white;
  padding: 0.5rem;
  border: none;
  border-radius: 0.25rem;
`;